+++ packages/riverpod_annotation/lib/experimental/mutation.dart
@@ -1,0 +1,15 @@
+// ignore_for_file: invalid_export_of_internal_element
+
+export '../src/internal.dart'
+    show
+        mutationZoneKey,
+        mutation,
+        Mutation,
+        MutationState,
+        IdleMutationState,
+        PendingMutationState,
+        ErrorMutationState,
+        SuccessMutationState,
+        MutationBase,
+        $SyncMutationBase,
+        $AsyncMutationBase;
