+++ packages/riverpod_generator/test/dependencies_test.dart
@@ -1,3 +1,4 @@
+import 'package:riverpod_annotation/experimental/scope.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:test/test.dart';
