+++ packages/riverpod_generator/test/integration/mutation.dart
@@ -1,3 +1,4 @@
+import 'package:riverpod_annotation/experimental/mutation.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'mutation.g.dart';
