+++ packages/riverpod_lint_flutter_test/test/lints/provider_dependencies/missing_dependencies.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:riverpod_annotation/experimental/scope.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'missing_dependencies.g.dart';
